+++ site/hugo_stats.json
@@ -81,7 +81,6 @@
       "-top-8",
       "-translate-x-1/2",
       "-translate-y-1/2",
-      "...",
       ".faq-acc-question-text",
       "2xl:bg-[position:50%_-50px]",
       "[box-shadow:0_4px_16px_rgba(0,_0,_0,_0.1)]",
@@ -204,11 +203,21 @@
       "flex-row",
       "flex-shrink-0",
       "flex-wrap",
+      "focus:absolute",
+      "focus:bg-[#ba0108]",
       "focus:bg-[#e3e3e3]",
       "focus:border-[#ba0108]",
       "focus:font-bold",
+      "focus:left-4",
+      "focus:no-underline",
       "focus:not-sr-only",
+      "focus:px-4",
+      "focus:py-2",
+      "focus:rounded",
       "focus:text-[#ba0108]",
+      "focus:text-white",
+      "focus:top-4",
+      "focus:z-50",
       "font-['Inter']",
       "font-['Montserrat']",
       "font-['Unbounded']",
