--- src/css/imports/_mobile-sliders.scss
@@ -1,75 +1,0 @@
-/* Mobile Slider Styles */
-#mobile-slider {
-  position: relative;
-  overflow: visible;
-}
-
-#mobile-slider-wrapper {
-  display: flex;
-  transition: tranform 0.3s ease-in-out;
-  width: 100%;
-}
-
-#mobile-slider-wrapper > div {
-  flex-shrink: 0;
-  width: 100%;
-  padding: 0 10px;
-}
-
-.mobile-prev,
-.mobile-next {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  z-index: 10;
-  cursor: pointer;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.2s ease;
-}
-
-.mobile-prev {
-  left: -3px;
-}
-
-.mobile-next {
-  right: -3px;
-}
-
-@media (min-width: 600px) and (max-width: 767px) {
-  .mobile-prev {
-    left: 80px !important;
-  }
-  
-  .mobile-next {
-    right: 80px !important;
-  }
-}
-
-@media (min-width: 476px) and (max-width: 599px) {
-  .mobile-prev {
-    left: 15px !important;
-  }
-  
-  .mobile-next {
-    right: 15px !important;
-  }
-}
-
-@media (max-width: 475px) {
-  .mobile-prev {
-    left: -3px !important;
-  }
-  
-  .mobile-next {
-    right: -3px !important;
-  }
-}
-
-.mobile-prev:hover,
-.mobile-next:hover {
-  transform: translateY(-50%) scale(1.05);
-}
