--- src/css/imports/_sliders-complete.scss
@@ -1,317 +1,0 @@
-/* Complete Slider Styles */
-
-/* Custom Slider Styles */
-.swiper-container-trusted,
-.swiper-container-products {
-  overflow: hidden;
-  position: relative;
-  height: auto;
-}
-
-.swiper-wrapper {
-  display: flex;
-  align-items: stretch;
-  height: auto;
-  transition: transform 0.5s cubic-bezier(0.25, 0.46, 0.45, 0.94);
-  flex-wrap: nowrap;
-  margin-left: 5px; 
-  margin-block: 10px;
-}
-
-.swiper-slide {
-  flex-shrink: 0;
-  display: flex;
-  align-items: center;
-  width: fit-content;
-  justify-content: center;
-}
-
-/* Products Slider Specific Styles */
-.swiper-container-products .swiper-slide {
-  width: auto !important; 
-  min-height: 400px;
-}
-
-/* Products slider responsive layout */
-.products-slider {
-  .swiper-container-products {
-    width: 100% !important;
-    max-width: 100% !important;
-    margin-inline: auto !important;
-    overflow: hidden !important; 
-  }
-  
-  @media (min-width: 768px) and (max-width: 860px) {
-    .swiper-container-products {
-      width: 85% !important;
-      max-width: 800px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 48% !important;
-      min-width: 320px !important;
-      max-width: 350px !important;
-    }
-  }
-
-  @media (min-width: 861px) and (max-width: 968px) {
-    .swiper-container-products {
-      width: 83% !important;
-      max-width: 950px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 48% !important;
-      min-width: 350px !important;
-      max-width: 380px !important;
-    }
-  }
-
-  @media (min-width: 969px) and (max-width: 1023px) {
-    .swiper-container-products {
-      width: 77% !important;
-      max-width: 900px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 48% !important;
-      min-width: 350px !important;
-      max-width: 380px !important;
-    }
-  }
-
-  @media (min-width: 1024px) and (max-width: 1099px) {
-    .swiper-container-products {
-      width: 76% !important;
-      max-width: 950px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 48% !important;
-      min-width: 380px !important;
-      max-width: 400px !important;
-    }
-  }
-  
-  /* Desktop styles - 2-3 slides visible depending on screen size */
-  @media (min-width: 1100px) and (max-width: 1199px) {
-    .swiper-container-products {
-      width: 72% !important;
-      max-width: 1000px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 48% !important;
-      min-width: 400px !important;
-      max-width: 400px !important;
-    }
-  }
-  
-  @media (min-width: 1200px) and (max-width: 1399px) {
-    .swiper-container-products {
-      width: 71% !important;
-      max-width: 1100px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 47% !important;
-      min-width: 400px !important;
-      max-width: 400px !important;
-    }
-  }
-  
-  @media (min-width: 1400px) {
-    .swiper-container-products {
-      width: 70% !important;
-      max-width: 1200px !important;
-    }
-    
-    .swiper-container-products .swiper-slide {
-      width: 45% !important;
-      min-width: 400px !important;
-      max-width: 400px !important;
-    }
-  }
-
-  .swiper-container-products .swiper-wrapper {
-    width: auto !important;
-    transition: transform 0.5s cubic-bezier(0.25, 0.46, 0.45, 0.94) !important;
-  }
-
-  /* Navigation arrows positioning */
-  .swiper-button-prev-custom,
-  .swiper-button-next-custom {
-    cursor: pointer;
-    transition: all 0.3s ease;
-  }
-
-  .swiper-button-prev-custom {
-    left: 0 !important;
-  }
-
-  .swiper-button-next-custom {
-    right: 0 !important;
-  }
-
-  @media (max-width: 1300px) {
-    .swiper-button-prev-custom {
-      left: 15px !important;
-    }
-    .swiper-button-next-custom {
-      right: 15px !important;
-    }
-  }
-
-  @media (max-width: 1100px) {
-    .swiper-button-prev-custom {
-      left: 15px !important;
-    }
-    .swiper-button-next-custom {
-      right: 15px !important;
-    }
-  }
-
-  @media (max-width: 900px) {
-    .swiper-button-prev-custom {
-      left: -15px !important;
-    }
-    .swiper-button-next-custom {
-      right: -15px !important;
-    }
-  }
-
-  @media (max-width: 800px) {
-    .swiper-button-prev-custom {
-      left: -5px !important;
-    }
-    .swiper-button-next-custom {
-      right: -5px !important;
-    }
-  }
-
-  /* Closer arrows for tablets */
-  @media (min-width: 768px) and (max-width: 1023px) {
-    .swiper-button-prev-custom {
-      left: 7px !important;
-    }
-    .swiper-button-next-custom {
-      right: 7px !important;
-    }
-  }
-}
-
-/* Trusted Slider Styles */
-.swiper-container-trusted-desktop .swiper-slide {
-  min-height: 200px;
-  width: 25% !important;
-  flex-shrink: 0;
-  padding: 0 20px;
-  max-width: 25% !important;
-  min-width: 25% !important;
-}
-
-.swiper-container-trusted-desktop .swiper-wrapper {
-  display: flex !important;
-  align-items: stretch !important;
-  height: auto !important;
-  transition: transform 0.5s cubic-bezier(0.25, 0.46, 0.45, 0.94) !important;
-  flex-wrap: nowrap !important;
-  width: 100% !important;
-  justify-content: space-between !important;
-}
-
-.swiper-container-trusted-mobile .swiper-slide {
-  min-height: auto;
-  width: 100% !important;
-  flex-shrink: 0;
-  padding: 0;
-  max-width: 100% !important;
-  min-width: 100% !important;
-}
-
-.swiper-pagination-trusted{
-  display: flex;
-  justify-content: center;
-  gap: 5px;
-  margin-bottom: 15px;
-}
-
-.swiper-pagination-trusted .swiper-pagination-bullet-custom {
-  width: 14px;
-  height: 14px;
-}
-
-.swiper-pagination-bullet-custom {
-  display: inline-block;
-  width: 14px;
-  height: 14px;
-  background: rgba(64, 64, 64, 1);
-  border-radius: 50%;
-  margin: 0 4px;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.swiper-pagination-bullet-active-custom {
-  background: #ba0108;
-}
-
-.swiper-pagination-trusted {
-  text-align: center;
-  margin-top: 20px;
-}
-
-@media (max-width: 450px) {
-    .swiper-pagination-trusted {
-      margin-top: 13px;
-    }
-}
-
-.swiper-container-trusted {
-  @media (min-width: 799px) {
-    .swiper-button-prev-custom {
-      position: absolute !important;
-      left: 100px !important;
-      top: 50% !important;
-      transform: translateY(-50%) !important;
-    }
-    
-    .swiper-button-next-custom {
-      position: absolute !important;
-      right: 100px !important;
-      top: 50% !important;
-      transform: translateY(-50%) !important;
-    }
-  }
-
-  @media (max-width: 798px) {
-    .swiper-button-prev-custom {
-      position: relative !important;
-      left: -3px !important;
-    }
-    
-    .swiper-button-next-custom {
-      position: relative !important;
-      right: -3px !important;
-    }
-  }
-}
-
-.swiper-container-trusted-desktop .swiper-wrapper {
-  display: flex;
-  align-items: stretch;
-  height: auto;
-  transition: none;
-  flex-wrap: nowrap;
-  width: 100% !important;
-  justify-content: space-between;
-}
-
-.swiper-container-trusted-mobile .swiper-wrapper {
-  display: flex;
-  align-items: stretch;
-  height: auto;
-  transition: none;
-  flex-wrap: nowrap;
-  width: 100% !important;
-}
